requirements.txt
discord.py
boto3
setuptools_scm
pyjokes
pillow
matplotlib
numpy
aiohttp

